refactor: icon 파일 styles 폴더로 이동
